+++ factory/src/main/java/io/github/factoryfx/factory/AttributeMetadataVisitor.java
@@ -1,0 +1,9 @@
+package io.github.factoryfx.factory;
+
+import io.github.factoryfx.factory.attribute.Attribute;
+import org.w3c.dom.Attr;
+
+@FunctionalInterface
+public interface AttributeMetadataVisitor {
+    void accept(String attributeVariableName, Class<? extends Attribute<?,?>> attributeClass, Class<? extends FactoryBase<?,?>> referenceClass);
+}